refactor: move paginated table
